refactor(cleanup): remove deprecated config fields, move datetime_utils, consolidate planner_logger
