+++ AGENTS.md
@@ -169,6 +169,7 @@ If credentials, API keys, or tokens are required:
 - Avoid introducing new dependencies unless justified and discussed with the user.
 - Apply Python style rules as defined in `pyproject.toml` and `tox.ini`.
 - Run `tox -e lint` locally before committing (runs isort, black, ruff format, ruff check).
+- Run `tox -e quality` after lint (runs pyright and vulture static checks).
 - See `CODE_QUALITY_STANDARDS.md` for full quality rules and conventions.
 - **Never use `==` or `!=` to compare floating-point values.** In production code use an epsilon
   guard (e.g. `abs(x) > 1e-9` instead of `x != 0`). In tests always use `pytest.approx()`.
@@ -254,6 +255,7 @@ Before creating a commit, the agent MUST report the result of:
 Before submitting a PR, the agent MUST:
 
 - Run `tox -e lint` to format and lint all code (runs isort, black, ruff format, and ruff check)
+- Run `tox -e quality` after lint (runs pyright and vulture static checks)
 - Run all tests locally: `pytest tests/`
 - Verify `git status` shows only intended changes
 - Commit changes with: `git commit -m "<type>(<scope>): <description>"`
@@ -330,9 +332,10 @@ A change is considered complete when:
 - New behavior is covered by tests (where feasible)
 - Code follows project style and conventions (enforced by isort, black, and ruff)
 - **All lint checks pass** (`tox -e lint` — runs isort, black, ruff format, ruff check)
+- **Quality checks pass** (`tox -e quality` — runs pyright and vulture)
 - Documentation is updated if configuration, API, or user-facing changes are made
 - No secrets are committed
-- All linting and formatting checks pass (`tox -e lint` and CI)
+- All linting, formatting, and quality checks pass (`tox -e lint` + `tox -e quality` and CI)
 - The implementation adheres strictly to the No-Assumption Rule
 - The change aligns with Home Assistant integration standards
 - Code quality is enhanced (no technical debt introduced)
