+++ custom_components/hsem/models/planner_inputs.py
@@ -90,7 +90,6 @@ class BatteryScheduleInput:
     enabled: bool = False
     start: time = time(0, 0)
     end: time = time(1, 0)
-    min_price_difference: float = 0.0
 
 
 @dataclass
@@ -127,9 +126,6 @@ class PlannerInput:
             Charge-side efficiency as a percentage (0-100).  Energy stored in
             the battery equals input energy × (charge_efficiency_pct / 100).
             Defaults to 95 % (5 % charge-side loss).
-        battery_conversion_loss_pct:
-            Round-trip conversion loss as a percentage (0-100).  Legacy field;
-            used by the cost function's conversion_loss_cost term.
         battery_discharge_efficiency_pct:
             Discharge-side efficiency as a percentage (0-100).  Energy delivered
             to the house equals battery energy removed × (discharge_efficiency_pct / 100).
@@ -191,7 +187,6 @@ class PlannerInput:
     battery_max_charge_power_w: float = 5000.0
     battery_max_discharge_power_w: float | None = None
     battery_charge_efficiency_pct: float = 95.0
-    battery_conversion_loss_pct: float = 10.0
     battery_discharge_efficiency_pct: float = 95.0
 
     # --- battery economics ---
