+++ custom_components/hsem/planner/candidate_selector.py
@@ -34,7 +34,6 @@ from __future__ import annotations
 import math
 from datetime import datetime, timedelta
 
-from custom_components.hsem.datetime_utils import as_tz
 from custom_components.hsem.models.planner_outputs import RejectedPlan
 from custom_components.hsem.planner.candidate_generator import (
     CANDIDATE_BASELINE,
@@ -44,6 +43,7 @@ from custom_components.hsem.planner.candidate_generator import (
 from custom_components.hsem.planner.cost_function import CostWeights, score_plan
 from custom_components.hsem.planner.planner_logger import log_planner
 from custom_components.hsem.planner.soc_simulation import simulate_soc
+from custom_components.hsem.utils.datetime_utils import as_tz
 from custom_components.hsem.utils.recommendations import Recommendations
 
 # SoC floor tolerance — plans are accepted even if they dip this many
