+++ custom_components/hsem/planner/slot_population.py
@@ -46,7 +46,6 @@ from custom_components.hsem.const import (
     SPIKE14_REDIST_TO_7D,
     SPIKE14_REDUCE_FRACTION_MAX,
 )
-from custom_components.hsem.datetime_utils import as_tz
 from custom_components.hsem.models.planner_inputs import (
     HourlyConsumptionAverage,
     PlannerInput,
@@ -55,6 +54,7 @@ from custom_components.hsem.models.planner_inputs import (
 )
 from custom_components.hsem.models.planner_outputs import PlannedSlot
 from custom_components.hsem.models.time_series import TimeSeriesIndex
+from custom_components.hsem.utils.datetime_utils import as_tz
 from custom_components.hsem.utils.prices import SlotPrice
 from custom_components.hsem.utils.recommendations import Recommendations
