--- custom_components/hsem/test_dummy.py
@@ -1,3 +1,0 @@
-# Test file for the HSEM integration
-def test_dummy():
-    pass
