+++ tests/planner/test_48h_second_day.py
@@ -111,13 +111,11 @@ def _make_48h_input(
             enabled=True,
             start=time(7, 0),
             end=time(9, 0),
-            min_price_difference=0.05,
         ),
         BatteryScheduleInput(
             enabled=True,
             start=time(17, 0),
             end=time(21, 0),
-            min_price_difference=0.05,
         ),
     ]
 
@@ -129,7 +127,6 @@ def _make_48h_input(
         battery_rated_capacity_kwh=10.0,
         battery_end_of_discharge_soc_pct=10.0,
         battery_max_charge_power_w=5000.0,
-        battery_conversion_loss_pct=10.0,
         battery_purchase_price=0.0,
         battery_expected_cycles=6000,
         weight_1d=25,
@@ -356,7 +353,6 @@ class TestDay2PreCharge:
                 enabled=True,
                 start=time(17, 0),
                 end=time(21, 0),
-                min_price_difference=0.10,
             )
         ]
 
@@ -368,7 +364,6 @@ class TestDay2PreCharge:
             battery_rated_capacity_kwh=10.0,
             battery_end_of_discharge_soc_pct=10.0,
             battery_max_charge_power_w=5000.0,
-            battery_conversion_loss_pct=0.0,
             battery_purchase_price=0.0,
             battery_expected_cycles=6000,
             weight_1d=25,
