+++ tests/planner/test_ev_planned_load.py
@@ -114,7 +114,6 @@ def _make_planner_input(
         battery_max_charge_power_w=5000.0,
         battery_max_discharge_power_w=5000.0,
         battery_charge_efficiency_pct=95.0,
-        battery_conversion_loss_pct=5.0,
         battery_discharge_efficiency_pct=95.0,
         weight_1d=25,
         weight_3d=30,
@@ -865,7 +864,6 @@ class TestEvSolarSurplusRegression:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=95.0,
-            battery_conversion_loss_pct=5.0,
             battery_discharge_efficiency_pct=95.0,
             weight_1d=25,
             weight_3d=30,
@@ -971,7 +969,6 @@ class TestEvSolarSurplusRegression:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=95.0,
-            battery_conversion_loss_pct=5.0,
             battery_discharge_efficiency_pct=95.0,
             weight_1d=25,
             weight_3d=30,
@@ -1203,7 +1200,6 @@ class TestEvAcLoadAndSoCPath:
             battery_end_of_discharge_soc_pct=5.0,
             battery_max_soc_pct=90.0,
             battery_max_charge_power_w=5000.0,
-            battery_conversion_loss_pct=0.0,
             battery_charge_efficiency_pct=100.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
@@ -1526,7 +1522,6 @@ class TestEvAcLoadAndSoCPath:
             battery_end_of_discharge_soc_pct=5.0,
             battery_max_soc_pct=90.0,
             battery_max_charge_power_w=5000.0,
-            battery_conversion_loss_pct=0.0,
             battery_charge_efficiency_pct=100.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
@@ -1706,7 +1701,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -1802,7 +1796,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -1904,7 +1897,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -1989,7 +1981,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -2079,7 +2070,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -2166,7 +2156,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -2244,7 +2233,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -2328,7 +2316,6 @@ class TestEvLoadSemantics:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
@@ -2491,8 +2478,7 @@ class TestEvLoadDoesNotInflateChargeNeeded:
         discharge_schedule = BatteryScheduleInput(
             enabled=True,
             start=_time(16, 0),
-            end=_time(22, 0),
-            min_price_difference=0.10,  # spread needed: 0.05 vs 0.80 → 0.75 > 0.10
+            end=_time(22, 0),  # spread needed: 0.05 vs 0.80 → 0.75 > 0.10
         )
 
         return PlannerInput(
@@ -2506,7 +2492,6 @@ class TestEvLoadDoesNotInflateChargeNeeded:
             battery_max_charge_power_w=5000.0,
             battery_max_discharge_power_w=5000.0,
             battery_charge_efficiency_pct=100.0,
-            battery_conversion_loss_pct=0.0,
             battery_discharge_efficiency_pct=100.0,
             weight_1d=25,
             weight_3d=30,
