+++ tests/planner/test_missing_tomorrow_data.py
@@ -185,7 +185,6 @@ def _make_48h_input(
         battery_end_of_discharge_soc_pct=10.0,
         battery_max_soc_pct=100.0,
         battery_max_charge_power_w=5000.0,
-        battery_conversion_loss_pct=10.0,
         battery_purchase_price=10_000.0,
         battery_expected_cycles=6000,
         weight_1d=25,
@@ -355,7 +354,6 @@ class TestCompleteTomorrowData:
             battery_soc_pct=50.0,
             battery_end_of_discharge_soc_pct=10.0,
             battery_max_charge_power_w=5000.0,
-            battery_conversion_loss_pct=10.0,
             consumption_averages=_consumption_averages(),
             price_points=_today_price_points(),
             solcast_slots=_pv_slots(),
@@ -683,7 +681,6 @@ class TestDataQualityAsDict:
             battery_soc_pct=50.0,
             battery_end_of_discharge_soc_pct=10.0,
             battery_max_charge_power_w=5000.0,
-            battery_conversion_loss_pct=10.0,
             consumption_averages=_consumption_averages(),
             price_points=_today_price_points(),
             solcast_slots=_pv_slots(),
