+++ tests/planner/test_multi_day_price_preservation.py
@@ -117,7 +117,6 @@ def _make_48h_input_with_day_offsets(
         battery_rated_capacity_kwh=10.0,
         battery_end_of_discharge_soc_pct=10.0,
         battery_max_charge_power_w=5000.0,
-        battery_conversion_loss_pct=10.0,
         battery_purchase_price=0.0,
         battery_expected_cycles=6000,
         weight_1d=25,
@@ -132,7 +131,6 @@ def _make_48h_input_with_day_offsets(
                 enabled=True,
                 start=time(7, 0),
                 end=time(9, 0),
-                min_price_difference=0.05,
             )
         ],
         excess_export_enabled=False,
@@ -379,7 +377,6 @@ class TestPlannerMultiDayPriceIsolation:
             battery_rated_capacity_kwh=10.0,
             battery_end_of_discharge_soc_pct=10.0,
             battery_max_charge_power_w=5000.0,
-            battery_conversion_loss_pct=10.0,
             battery_purchase_price=0.0,
             battery_expected_cycles=6000,
             weight_1d=25,
