+++ tests/planner/test_planning_horizon.py
@@ -135,13 +135,11 @@ def _make_input(
             enabled=True,
             start=time(7, 0),
             end=time(9, 0),
-            min_price_difference=0.05,
         ),
         BatteryScheduleInput(
             enabled=True,
             start=time(17, 0),
             end=time(21, 0),
-            min_price_difference=0.05,
         ),
     ]
     return PlannerInput(
@@ -153,7 +151,6 @@ def _make_input(
         battery_end_of_discharge_soc_pct=10.0,
         battery_max_soc_pct=100.0,
         battery_max_charge_power_w=5000.0,
-        battery_conversion_loss_pct=10.0,
         battery_purchase_price=0.0,
         battery_expected_cycles=6000,
         weight_1d=25,
