+++ tests/test_dst_transitions.py
@@ -339,7 +339,6 @@ class TestPlannerRunsDstDays:
                 enabled=True,
                 start=time(17, 0),
                 end=time(21, 0),
-                min_price_difference=0.05,
             )
         ]
         return PlannerInput(
@@ -350,7 +349,6 @@ class TestPlannerRunsDstDays:
             battery_rated_capacity_kwh=10.0,
             battery_end_of_discharge_soc_pct=10.0,
             battery_max_charge_power_w=5000.0,
-            battery_conversion_loss_pct=10.0,
             battery_purchase_price=10_000.0,
             battery_expected_cycles=6000,
             weight_1d=25,
